Refactoring to support better and consistent styling
