+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/VariableStarViewer/components/Controls/Controls.js
@@ -1,11 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import {
-  Box,
-  FormField,
-  RadioButtonGroup
-} from 'grommet'
-import styled, { css } from 'styled-components'
+import { Box } from 'grommet'
+import styled from 'styled-components'
 import { PlainButton, SpacedText, withThemeContext } from '@zooniverse/react-components'
 import counterpart from 'counterpart'
 import FlipIcon from '../FlipIcon'
@@ -13,6 +9,7 @@ import en from '../../locales/en'
 import theme from './theme'
 import VisibilitySeriesCheckBoxes from './components/VisibilitySeriesCheckBoxes'
 import PhaseFocusControls from './components/PhaseFocusControls'
+import PeriodMultipleControls from './components/PeriodMultipleControls'
 
 counterpart.registerTranslations('en', en)
 
@@ -30,28 +27,10 @@ export const FlipButton = styled(PlainButton)`
   }
 `
 
-export const StyledRadioButtonGroup = styled(RadioButtonGroup)`
-  > div {
-    position: relative;
-  }
-
-  > div:after {
-    ${props => css`border-bottom: 1px solid ${props.color};`}
-    bottom: 0;
-    content: '';
-    height: 5px;
-    left: 1px;
-    position: absolute;
-    right: 0;
-    top: 0;
-    width: 28px;
-  }
-`
 
 function Controls(props) {
   const {
     data,
-    visibleSeries,
     gridArea,
     periodMultiple,
     periodMultipleOptions,
@@ -60,7 +39,7 @@ function Controls(props) {
     setSeriesPhaseFocus,
     setPeriodMultiple,
     setYAxisInversion,
-    theme
+    visibleSeries
   } = props
 
   const radioButtonOptions = periodMultipleOptions.map((option) => {
@@ -74,12 +53,12 @@ function Controls(props) {
     <Box
       background='neutral-6'
       direction='row'
-      gap='xsmall'
+      justify='between'
       gridArea={gridArea}
       pad='xsmall'
     >
       <SpacedText
-        style={{ fontSize: '0.5em', transform: 'rotate(180deg)', writingMode: 'vertical-lr' }}
+        style={{ fontSize: '0.5em', textAlign: 'center', transform: 'rotate(180deg)', writingMode: 'vertical-lr' }}
       >
         {counterpart('VariableStarViewer.controls')}
       </SpacedText>
@@ -93,23 +72,12 @@ function Controls(props) {
         onClick={event => setYAxisInversion(event)}
         pad='small'
       />
-      <FormField
-        htmlFor='periodMultiple'
-        label={<SpacedText size='10px' weight='bold'>{counterpart('VariableStarViewer.periodMultiple')}</SpacedText>}
-        style={{ position: 'relative' }}
-      >
-        <StyledRadioButtonGroup
-          color={theme.global.colors['light-6']}
-          direction='row'
-          gap='medium'
-          id='periodMultiple'
-          name='periodMultiple'
-          onChange={event => setPeriodMultiple(event)}
-          options={radioButtonOptions}
-          value={periodMultiple.toString()}
-        />
-      </FormField>
-      <Box>
+      <PeriodMultipleControls
+        options={radioButtonOptions}
+        periodMultiple={periodMultiple}
+        setPeriodMultiple={setPeriodMultiple}
+      />
+      <Box justify='between'>
         <PhaseFocusControls
           data={data}
           phaseFocusedSeries={phaseFocusedSeries}
@@ -125,7 +93,7 @@ function Controls(props) {
           visibleSeries={visibleSeries}
           setSeriesVisibility={setSeriesVisibility}
         />
-        <SpacedText color='black' size='10px' weight='bold'>
+        <SpacedText size='10px' weight='bold'>
           {counterpart('VariableStarViewer.visibility')}
         </SpacedText>
       </Box>
@@ -141,11 +109,6 @@ Controls.defaultProps = {
   setSeriesVisibility: () => {},
   setPeriodMultiple: () => {},
   setYAxisInversion: () => {},
-  theme: {
-    global: {
-      colors: {}
-    }
-  },
   visibleSeries: []
 }
 
@@ -156,7 +119,6 @@ Controls.propTypes = {
   setSeriesVisibility: PropTypes.func,
   setPeriodMultiple: PropTypes.func,
   setYAxisInversion: PropTypes.func,
-  theme: PropTypes.object,
   visibleSeries: PropTypes.arrayOf(PropTypes.object)
 }
