+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/VariableStarViewer/components/Controls/components/Label/index.js
@@ -1,0 +1,1 @@
+export { default } from './Label'
