+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/VariableStarViewer/components/Controls/components/PeriodMultipleControls/index.js
@@ -1,0 +1,1 @@
+export { default } from './PeriodMultipleControls'
