+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/VariableStarViewer/components/Controls/components/PeriodMultipleControls/theme.js
@@ -2,20 +2,6 @@ import { css } from 'styled-components'
 
 const theme = {
   checkBox: {
-    border: {
-      color: {
-        light: 'light-6',
-        dark: 'light-6' // this will change when dark theme styles is added
-      }
-    },
-    hover: {
-      border: {
-        color: {
-          dark: 'brand',
-          light: 'brand'
-        }
-      }
-    },
     icon: {
       size: '20px'
     },
@@ -35,9 +21,22 @@ const theme = {
     border: {
       color: {
         light: 'light-6',
-        dark:  'light-6' // this will change when dark theme styles is added
+        dark: 'light-6' // this will change when dark theme styles is added
       }
     },
+    extend: () => css`
+      flex-direction: column;
+      position: relative;
+      z-index: 100;
+
+      input:not(:checked) + div {
+        height: 5px;
+        margin-bottom: 5px;
+        position: relative;
+        top: 2.5px;
+        width: 5px;
+      }
+    `,
     hover: {
       border: {
         color: {
@@ -49,7 +48,7 @@ const theme = {
     gap: '0px',
     size: '10px',
     icon: {
-      size: '16px'
+      size: '6px'
     }
   }
 }
