--- packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/VariableStarViewer/components/Controls/components/StyledLabel/index.js
@@ -1,1 +1,0 @@
-export { default } from './StyledLabel'
