+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/VariableStarViewer/components/Controls/components/VisibilitySeriesCheckBoxes/VisibilitySeriesCheckBoxes.js
@@ -1,10 +1,10 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { Box } from 'grommet'
+import { Box, CheckBox } from 'grommet'
 import styled, { css, withTheme } from 'styled-components'
 import { transparentize } from 'polished'
 import { SpacedText } from '@zooniverse/react-components'
-import StyledLabel from '../StyledLabel'
+import Label from '../Label'
 import getDataSeriesColor from '../../../../../../helpers/getDataSeriesColor'
 import getDataSeriesSymbol from '../../../../../../helpers/getDataSeriesSymbol'
 
@@ -21,43 +21,30 @@ function VisibilitySeriesCheckBoxes (props) {
   } = props
 
   return (
-    <Box direction='row' gap='xsmall' pad='none'>
+    <Box direction='column' pad='none'>
       {visibleSeries.map((series, seriesIndex) => {
         const [[label, checked]] = Object.entries(series)
         const seriesOptions = data[seriesIndex]?.seriesOptions
-        const color = getDataSeriesColor({
-          defaultColors: Object.values(colors.drawingTools),
-          visibleSeries,
-          seriesOptions: seriesOptions,
-          seriesIndex,
-          themeColors: colors
-        })
-        const Glyph = getDataSeriesSymbol(seriesIndex)
+
         return (
-          <StyledLabel
-            borderColor={(checked) ? colors['light-6'] : transparentize(0.5, colors['light-6'])}
-            htmlFor={label}
+          <CheckBox
+            checked={checked}
             key={`${label}-${seriesIndex}`}
-          >
-            <input
-              checked={checked}
-              id={label}
-              name='series-focus'
-              onChange={event => { setSeriesVisibility(event) }}
-              type='checkbox'
-              value={label}
-            />
-            <svg viewBox='0 0 10 10' width='15px' style={{ paddingRight: '0.5ch' }}>
-              <Glyph left={5} fill={color} size={20} top={5} />
-            </svg>
-            <SpacedText
-              color={(checked) ? 'black' : 'rbga(0,0,0,0.5)'}
-              style={{ fontSize: '0.5em', whiteSpace: 'nowrap' }} // TODO: update theme for smaller size of span text
-              weight='bold'
-            >
-              {label}
-            </SpacedText>
-          </StyledLabel>
+            label={
+              <Label
+                colors={colors}
+                checked={checked}
+                label={label}
+                seriesIndex={seriesIndex}
+                seriesOptions={seriesOptions}
+                visibleSeries={visibleSeries}
+              />
+            }
+            name='series-visibility'
+            onChange={event => { setSeriesVisibility(event) }}
+            type='checkbox'
+            value={label}
+          />
         )
       })}
     </Box>
